ヘッドショット時と胴体を攻撃した時にダメージが変化するようにしました
