+++ Assets/Resources/PlayerPrefab.prefab
@@ -267,6 +267,7 @@ GameObject:
   - component: {fileID: 114960514103121954}
   - component: {fileID: 114677642978999624}
   - component: {fileID: 114634925773028886}
+  - component: {fileID: -8141770205055736597}
   m_Layer: 9
   m_Name: PlayerPrefab
   m_TagString: Player
@@ -430,6 +431,7 @@ MonoBehaviour:
     serializedVersion: 2
     m_Bits: 513
   speed: 6
+  speedUp: 0
   jumpSpeed: 1.5
   gravity: 2.5
   minAngle: -30
@@ -485,7 +487,22 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1eddbf297682bc54fab7971f90912d07, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_initialLife: 1000
+  m_manager: {fileID: -8141770205055736597}
+  m_damage: 10
+  m_photonView: {fileID: 114960514103121954}
+--- !u!114 &-8141770205055736597
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2871077469082834734}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 42ceb983785429d43b71733fd81b6f21, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_initialLife: 5000
 --- !u!1 &2871077469509734706
 GameObject:
   m_ObjectHideFlags: 0
@@ -661,6 +678,7 @@ GameObject:
   m_Component:
   - component: {fileID: 3367170921607849059}
   - component: {fileID: 2075622815357316241}
+  - component: {fileID: 4301907556807190770}
   m_Layer: 9
   m_Name: hedColleder
   m_TagString: Hed
@@ -690,11 +708,26 @@ SphereCollider:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3352816202681330405}
   m_Material: {fileID: 0}
-  m_IsTrigger: 0
+  m_IsTrigger: 1
   m_Enabled: 1
   serializedVersion: 2
   m_Radius: 0.5
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!114 &4301907556807190770
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3352816202681330405}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1eddbf297682bc54fab7971f90912d07, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_manager: {fileID: -8141770205055736597}
+  m_damage: 100
+  m_photonView: {fileID: 114960514103121954}
 --- !u!1 &3494194190902498607
 GameObject:
   m_ObjectHideFlags: 0
