+++ Assets/Scripts/HPManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 42ceb983785429d43b71733fd81b6f21
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
